Started work on responsive design mode
Everything is being rendered in the correct place, but the collisions are all scaled wrong.
I don't know why that is, because it should all use the same exact inputs.

+++ Player.pde
@@ -97,7 +97,7 @@ class Player extends EntitySprite {
     }
   }
 
-  void changeSize(int newSize) {
+  void changeSize(float newSize) {
     super.changeSize(newSize);
     sword.resize(int(size.x), int(size.y));
   }
